GROOV-1929

+++ lib/groovepacker/stores/importers/shipstation_rest/orders_importer.rb
@@ -263,7 +263,7 @@ module Groovepacker
 
             def find_or_init_new_order(order)
               shipstation_order = Order.find_by_store_id_and_increment_id(@credential.store_id, order["orderNumber"])
-              return if shipstation_order && (shipstation_order.status=="scanned" || shipstation_order.status=="cancelled")
+              return if shipstation_order && (shipstation_order.status=="scanned" || shipstation_order.status=="cancelled" || shipstation_order.order_items.map(&:scanned_status).include?("partially_scanned") || shipstation_order.order_items.map(&:scanned_status).include?("scanned"))
               if @import_item.import_type == 'quick' && shipstation_order
                 shipstation_order.destroy
                 shipstation_order = nil
